# internal/ttto/repository/storage/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.18.0

package storage

import (
	"database/sql"
	"encoding/json"
)

type Game struct {
	ID              int32
	OwnerID         int32
	OpponentID      sql.NullInt32
	CurrentPlayerID sql.NullInt32
	StepCount       int32
	WinnerID        sql.NullInt32
	Field           json.RawMessage
	CurrentState    string
}

type GamesHistory struct {
	ID              int32
	GameID          int32
	OwnerID         int32
	OpponentID      sql.NullInt32
	CurrentPlayerID sql.NullInt32
	StepCount       int32
	WinnerID        sql.NullInt32
	Field           json.RawMessage
	CurrentState    string
}

type GamesWithUsername struct {
	ID                int32
	OwnerID           int32
	OpponentID        sql.NullInt32
	CurrentPlayerID   sql.NullInt32
	StepCount         int32
	WinnerID          sql.NullInt32
	Field             json.RawMessage
	CurrentState      string
	OwnerName         sql.NullString
	OpponentName      sql.NullString
	CurrentPlayerName sql.NullString
	WinnerName        sql.NullString
}

type User struct {
	ID       int32
	Username string
	Password string
}